Move common types and consts to separate file
