+++ src/index.ts
@@ -1,3 +1,4 @@
+export * from './common/app-meta';
 export * from './common/consts';
 export * from './endpoints/app-details';
 export * from './endpoints/top-charts';
